+++ src/Products/MeetingCommunes/profiles/examples_fr/import_data.py
@@ -859,8 +859,9 @@ data.directory_position_types = [
     {'token': u'secretaire',
      'name': u'Secrétaire de séance|Secrétaires de séance|Secrétaire de séance|Secrétaires de séance'},
 ]
-contactsTemplate = PodTemplateDescriptor('contactsTemplate', 'Export organizations', dashboard=True)
+contactsTemplate = PodTemplateDescriptor('contactsTemplate', 'Export', dashboard=True)
 contactsTemplate.odt_file = 'organizations-export.ods'
+contactsTemplate.use_objects = True
 contactsTemplate.pod_formats = ['ods', 'xls', ]
 contactsTemplate.dashboard_collections_ids = ['all_orgs']
 data.contactsTemplates = [contactsTemplate]