+++ CHANGES.rst
@@ -3,7 +3,7 @@ Products.MeetingCommunes Changelog
 
 The Products.MeetingCommunes version must be the same as the Products.PloneMeeting version
 
-4.1.17 (unreleased)
+4.1.17 (2020-10-01)
 -------------------
 
 - Backport from 4.2.x: